cleaning up romulus index pages with more components
