+++ src/pages/romulus/[address]/index.tsx
@@ -1,17 +1,13 @@
 import { useContractKit } from "@celo-tools/use-contractkit";
-import { BigNumber } from "ethers";
 import { useRouter } from "next/dist/client/router";
 import React from "react";
-import { Box, Button, Flex, Heading, Text, Image } from "theme-ui";
-
+import { Box, Button, Flex, Heading, Text } from "theme-ui";
 import { useDelegateModal } from "../../../components/pages/romulus/delegateModal";
-import { ProposalCard } from "../../../components/pages/romulus/ProposalCard";
 import { TopDelegates } from "../../../components/pages/romulus/TopDelegates";
 import {
   PoofToken__factory,
   RomulusDelegate__factory,
 } from "../../../generated";
-import { useProposals } from "../../../hooks/romulus/useProposals";
 import { useRomulus } from "../../../hooks/romulus/useRomulus";
 import { useVotingTokens } from "../../../hooks/romulus/useVotingTokens";
 import { useLatestBlockNumber } from "../../../hooks/useLatestBlockNumber";
@@ -23,12 +19,11 @@ import { truncateAddress } from "../../../util/address";
 import { BIG_ZERO } from "../../../util/constants";
 import { humanFriendlyWei } from "../../../util/number";
 import { governanceLookup } from "..";
-import styled from "styled-components";
 import AppBody from "../../AppBody";
-import { TopSection, AutoColumn } from "../../../components/Column";
-import { RowFlat, RowBetween, Row } from "../../../components/Row";
-import Loader from "../../../components/Loader";
-import { ProtocolImage } from "../../../components/Image";
+import { RowFlat } from "../../../components/Row";
+import { OverviewCard } from "../../../components/pages/romulus/OverviewCard";
+import { ProposalList } from "../../../components/pages/romulus/ProposalList";
+import { DetailContainer } from "../../../components/pages/romulus/Header";
 
 const RomulusIndexPage: React.FC = () => {
   const router = useRouter();
@@ -39,7 +34,6 @@ const RomulusIndexPage: React.FC = () => {
     ? governanceLookup[romulusAddress.toString()]
     : undefined;
   const { address } = useContractKit();
-  const [proposals] = useProposals((romulusAddress as string) || "");
   const [
     [
       hasReleaseToken,
@@ -117,30 +111,9 @@ const RomulusIndexPage: React.FC = () => {
   return (
     <>
       <Box>
-        <TopSection gap="md">
-          <DataCard>
-            <CardSection>
-              <AutoColumn gap="md">
-                <Row>
-                  {governanceDescription && (
-                    <ProtocolImage src={governanceDescription.icon} />
-                  )}
-
-                  <Text sx={{ fontWeight: 600 }}>
-                    {governanceDescription ? governanceDescription.name : ""}{" "}
-                    Governance Overview
-                  </Text>
-                </Row>
-                <RowBetween>
-                  <Text sx={{ fontSize: 14 }}>
-                    Create and view proposals, delegate votes, and participate
-                    in protocol governance!{" "}
-                  </Text>
-                </RowBetween>{" "}
-              </AutoColumn>
-            </CardSection>
-          </DataCard>
-        </TopSection>
+        <OverviewCard
+          governanceDescription={governanceDescription}
+        ></OverviewCard>
         <AppBody>
           <Box mb={4}>
             <RowFlat>
@@ -231,108 +204,17 @@ const RomulusIndexPage: React.FC = () => {
           </Box>
         </AppBody>
 
-        <AppBody>
-          <CreateProposalContainer>
-            <Heading as="h2" style={{ fontSize: "1.5rem" }}>
-              Governance Proposals
-            </Heading>
-            {!totalVotingPower.lt(BigNumber.from(proposalThreshold)) && (
-              <Button
-                onClick={() => {
-                  if (romulusAddress) {
-                    router
-                      .push(`/romulus/${romulusAddress.toString()}/create`)
-                      .catch(console.error);
-                  }
-                }}
-                disabled={totalVotingPower.lt(
-                  BigNumber.from(proposalThreshold)
-                )}
-              >
-                Create Proposal
-              </Button>
-            )}
-          </CreateProposalContainer>
-          <Box style={{ paddingBottom: "15px" }}>
-            {proposals.length > 1 ? (
-              proposals
-                .slice(-3)
-                .reverse()
-                .map((proposalEvent, idx) => (
-                  <Box key={idx} mt={3} style={{ margin: "32px" }}>
-                    <ProposalCard
-                      proposalEvent={proposalEvent}
-                      clickable={false}
-                      showId={true}
-                      showAuthor={false}
-                    />
-                  </Box>
-                ))
-            ) : (
-              <Loader size="48px"></Loader>
-            )}
-            <Box style={{ margin: "32px" }}>
-              <Button
-                onClick={() => {
-                  if (romulusAddress) {
-                    router
-                      .push(`/romulus/${romulusAddress.toString()}/proposals`)
-                      .catch(console.error);
-                  }
-                }}
-                style={{ width: "100%" }}
-              >
-                View more proposals
-              </Button>
-            </Box>
-          </Box>
-        </AppBody>
+        <ProposalList
+          proposalThreshold={proposalThreshold}
+          totalVotingPower={totalVotingPower}
+        ></ProposalList>
 
-        <AppBody>
-          <Box mb={4} sx={{ margin: "25px 15px 32px 15px", padding: "25px" }}>
-            <Heading as="h2" mb={3} style={{ fontSize: "1.25rem" }}>
-              Top delegates
-            </Heading>
-            <TopDelegates romulusAddress={romulusAddress as string} />
-          </Box>
-        </AppBody>
+        <TopDelegates romulusAddress={romulusAddress as string} />
       </Box>
       {tokenDelegateModal}
       {releaseTokenDelegateModal}
     </>
   );
 };
-
-export const DataCard = styled(AutoColumn)<{ disabled?: boolean }>`
-  background: radial-gradient(
-    96.02% 99.41% at 1.84% 0%,
-    ${(props) => props.theme.primary1} 30%,
-    ${(props) => props.theme.bg5} 100%
-  );
-  border-radius: 12px;
-  width: 100%;
-  position: relative;
-  overflow: hidden;
-`;
-
-export const CardSection = styled(AutoColumn)<{ disabled?: boolean }>`
-  padding: 1.5rem;
-  opacity: ${({ disabled }) => disabled && "0.4"};
-`;
-
-export const CreateProposalContainer = styled(Box)`
-  display: flex;
-  justify-content: space-between;
-  padding: 45px 45px 5px 45px;
-`;
-
-const DetailContainer = styled(Box)`
-  border: 3px solid #6d619a;
-  border-radius: 8px;
-  padding: 15px;
-  margin-bottom: 16px;
-  height: 150px;
-  width: 350px;
-`;
 
 export default RomulusIndexPage;
